error 修正
11月19日

+++ app/Http/Controllers/DiaryController.php
@@ -41,7 +41,7 @@ class DiaryController extends Controller
         $diary = new Diary();
 
         //Diaryモデルを使って、DBに日記を保存
-       //$diary->カラム名 = カラム　に設定したい値
+       //$diary->カラム名 = カラム に設定したい値
         $diary->title = $request->title;
         $diary->body = $request->body;
 
@@ -60,6 +60,11 @@ class DiaryController extends Controller
 
         //diaryモデルを使用してIDが一致する日記の取得
         $diary = Diary::find($id);
+        //ログインユーザーが日記の投稿者かチェックする
+    if(Auth::user()->id != $diary->user_id) {
+
+        abort(403);
+    }
         //取得した日記の削除
         $diary->delete();
         //一覧画面にリダイレクト
@@ -67,10 +72,15 @@ class DiaryController extends Controller
     }
 
     //編集画面を表示する
-    public function edit(int $id)
+    public function edit(diary $diary)
     {
+        //ログインユーザーが日記の投稿者かチェックする
+        if(Auth::user()->id != $diary->user_id) {
+
+            abort(403);
+        }
         // 受け取ったIDを元に日記を取得する
-        $diary = Diary::find($id);
+        // $diary = Diary::find($id);
         //編集画面を返す。同時に画面に取得した日記を渡す。
 
         return view('diaries.edit',[
@@ -86,6 +96,12 @@ class DiaryController extends Controller
     {
     //受け取ったIDを元に日記を取得
     $diary = Diary::find($id);
+
+    //ログインユーザーが日記の投稿者かチェックする
+    if(Auth::user()->id != $diary->user_id) {
+
+        abort(403);
+    }
     //取得した日記のタイトル、本文を書き換える
     //Diaryモデルを使って、DBに日記を保存
        //$diary->カラム名 = カラム　に設定したい値
